requirements.txt
mcp
pyyaml
httpx
gitpython